// src/models/user/IUser.ts
import type {ICrypto} from "./ICrypto.ts";
import type {IAddress} from "./IAddress.ts";
import type {IBank} from "./IBank.ts";
import type {ICompany} from "./ICompany.ts";

export interface IUser {
    id: number;
    firstName: string;
    lastName: string;
    maidenName: string;
    age: number;
    gender: string;
    email: string;
    phone: string;
    username: string;
    password: string;
    birthDate: string;
    image: string;
    bloodGroup: string;
    height: number;
    weight: number;
    eyeColor: string;
    hair: IHair;
    ip: string;
    address: IAddress;
    macAddress: string;
    university: string;
    bank: IBank;
    company: ICompany;
    ein: string;
    ssn: string;
    userAgent: string;
    crypto: ICrypto;
    role: string;
}


export interface IHair {
    color: string;
    type: string;
}